requirements.txt
fastapi==0.65.1
uvicorn==0.13.4
SQLAlchemy==1.4.15
psycopg2==2.8.6
alembic==1.6.3
python-dotenv==0.17.1
python-jose==3.2.0
passlib==1.7.4
PyJWT==2.1.0
passlib==1.7.4